+++ tools/monitor/ui/src/main/java/is/codion/tools/monitor/ui/ClientInstanceMonitorPanel.java
@@ -158,7 +158,7 @@ public final class ClientInstanceMonitorPanel extends JPanel {
 						.control(Control.builder()
 										.toggle(lineWrapState)
 										.name("Line wrap")))
-						.createPopupMenu());
+						.buildPopupMenu());
 
 		return textArea;
 	}