+++ changelog.md
@@ -6,6 +6,7 @@ Codion Change Log
 - PropertyStore, property access now based on predicates instead of key prefixes.
 ### is.codion.swing.common.ui
 - FilterTable, now propagates the associated MouseEvent as the source of the ActionEvent when doubleClickAction is performed.
+- MenuBuilder no longer extends MenuItemBuilder.
 ### is.codion.swing.framework.ui
 - EntityTablePanel bug fixed, Config.SHOW_REFRESH_PROGRESS_BAR now used for default config value.
 - EntityDialogs bug fixed, now respects the location if specified.